v1.23 message read fix
In the case you have sent a message, it is unseen and you reload the app, now it detect and log it correct when your friend your message has seen.
It takes also the time from label when you wrote the message. So the time until friend has seen will now correct logged.